checkpoint refactor
