+++ src/extract.py
@@ -11,33 +11,7 @@ import urllib.parse
 from github_api import github_readme_text
 from pdf_text import pdf_text
 
-
-class ExtractException(Exception):
-    """
-    various extraction errors
-    """
-    
-class UnsupportedHostException(ExtractException):
-    """
-    The URL is for a host we know we can't access reliably
-    """
-
-class UnsupportedContentType(ExtractException):
-    """
-    The http content type is unsupported.
-    """
-    
-class EmptyText(ExtractException):
-    """
-    Unable to extract any readable text from the URL.
-    """
-
-
-class NetworkError(ExtractException):
-    """
-    Unable to access the content.
-    """
-
+from exceptions import *
     
 
 def extract_text_from_html(content):
